requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
chromadb>=0.4.0
tiktoken>=0.5.0
pypdf>=3.0.0
pyyaml>=6.0